+++ src/earthkit/time/__init__.py
@@ -1,4 +1,4 @@
-from .climatology import date_range, model_climate_dates
+from .climatology import RelativeYear, date_range, model_climate_dates
 from .sequence import (
     DailySequence,
     MonthlySequence,
@@ -7,7 +7,11 @@ from .sequence import (
     YearlySequence,
 )
 
+__version__ = "0.1.7"
+
 __all__ = [
+    "__version__",
+    "RelativeYear",
     "date_range",
     "model_climate_dates",
     "DailySequence",