+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.1.7 - 2025-01-02
+
+* Export version as `earthkit.time.__version__`
+* Add options for relative year range in climatology tools
+
 ## 0.1.6 - 2024-11-04
 
 * Update project metadata